+++ doc/release_notes.rst
@@ -1,9 +1,13 @@
 Release Notes
 =============
 
-.. Upcoming Release
-.. ----------------
+Upcoming Release
+----------------
 
+
+* New solver interface for `SCIP <https://www.scipopt.org/>`. This solver is now supported by `linopy` and can be used with the `solve` function if the `pyscipopt` package is installed. The solver is available for free for general use. See the `SCIP website <https://www.scipopt.org/>` for more information.
+* Linopy was refactored to use the new xarray API (>=2024.01) without the deprecation warnings.
+* The set "quadratic_solvers" now only contains quadratic solvers which are installed and available to the user.
 
 Version 0.3.2
 -------------
@@ -15,6 +19,7 @@ Version 0.3.2
 
 Version 0.3.1
 -------------
+
 
 **New Features**
 
@@ -47,6 +52,7 @@ Version 0.2.6
 * The memory-efficiency of the IO to LP/MPS file was further improved. In particular, the function `to_dataframe` is now avoiding unnecessary data copies.
 * The printout of time stamps was modified to be more readable, leaving out the display of seconds and below if not necessary.
 * The gurobi environment is now enclosed in a context manager to avoid any unwanted use of a token.
+
 
 Version 0.2.5
 -------------